src/utils/tts/requirements.txt
#Comms
aiomqtt

#TTS
onnxruntime==1.17.0
piper-tts